requirements.txt
Pillow
zope.event
dropbox
mock
unittest-dataprovider
flake8
coverage